IrinaMinasyan/requirements.txt
requests
bs4
html2text
pytest
dateutil
dominate
fpdf2